fix(neo-n3): correct on-chain native calls + VM/NEF fidelity vs live mainnet
Neo N3 platform-correctness pass, verified against the live mainnet
native ABIs (getnativecontracts) and neo-project/neo source. (All 11
native contract hashes — including the v3.9.0 Treasury native — were
independently re-confirmed correct against mainnet and are unchanged.)

On-chain correctness (emitted artifacts):
- blockhash(n): was lowered to a non-existent Ledger.getBlockHash native
  method — it compiled clean but FAULTED method-not-found on a real node
  for every use (blockhash is advertised as EVM auto-compat). Now lowers
  to the real Ledger.getBlock(n).hash (Block field 0), null-guarded to
  return bytes32(0) for an unavailable/out-of-range index (matching EVM).
  The emitted manifest now grants the real  permission.
- Removed the fabricated Ledger.getBlockHash and Ledger.getBlockSystemFee
  entries from the native registry + resolve whitelist (neither exists on
  the live LedgerContract ABI; both would fault on a real node). Neither
  was referenced by any devpack .sol or test.
- NEF method-token cap: 512 -> 128. A real node hard-caps token
  deserialization at 128 (NefFile.cs ReadSerializableArray<MethodToken>(128)),
  so a NEF with >128 tokens was accepted locally but rejected on deploy.

In-tree NeoVM simulator fidelity (test-only; no emitted-bytecode change):
- System.Runtime.GetTrigger returned 0x10; the real Application trigger is
  0x40 (TriggerType: OnPersist 0x01, PostPersist 0x02, Verification 0x20,
  Application 0x40).
- ExecutionEngineLimits.MaxItemSize was 65535; the real value is
  ushort.MaxValue*2 = 131070 (verified in neo-vm ExecutionEngineLimits.cs).

Verified in a clean worktree at HEAD (isolated from unrelated in-tree
refactoring WIP): lib + e2e + runtime 548 passed / 0 failed; blockhash
runtime probe returns 0-or-32 bytes without faulting; token-cap and
blockhash fuzz harnesses green.

+++ src/runtime/execution/syscalls/runtime.rs
@@ -8,9 +8,11 @@ impl ExecutionContext {
                 Ok(true)
             }
             "System.Runtime.GetTrigger" => {
-                // Default to Application trigger (0x10) for the embedded runtime.
-                // This matches the most common execution context for deployed contracts.
-                self.push_stack(StackItem::Integer(0x10))?;
+                // Application trigger = 0x40 in Neo N3's TriggerType enum
+                // (OnPersist 0x01, PostPersist 0x02, Verification 0x20,
+                // Application 0x40). The embedded runtime models the common
+                // deployed-contract case, which executes under Application.
+                self.push_stack(StackItem::Integer(0x40))?;
                 Ok(true)
             }
             "System.Runtime.Platform" => {
